requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
pillow==10.0.1
moviepy==1.0.3
fer==22.1.0
mediapipe==0.10.7
matplotlib==3.7.2
numpy==1.24.3
requests==2.31.0
openai==1.3.0
transformers==4.35.0
torch==2.1.0
plotly==5.17.0
streamlit-option-menu==0.3.6
streamlit-camera-input-live==0.1.0


